requirements.txt
Appium-Python-Client==1.1.0
selenium==3.141.0
urllib3==1.25.6
astroid==2.2.5
behave==1.2.6
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
parse==1.12.1
parse-type==0.5.2
pylint==2.3.1
six==1.12.0
typed-ast==1.4.2
wrapt==1.11.2
requests==2.22.0
python-dotenv==0.10.3
ipdb==0.10.3
cffi==1.13.2
cryptography==2.8
pycparser==2.19
pyOpenSSL==19.1.0
DateTime==4.3
zope.interface==5.3.0
